requirements.txt
fastapi
python-multipart
twilio
sqlalchemy
langchain
langchain-openai
python-dotenv
uvicorn
